Component testing/Zhakhangir/requirements.txt
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.12.2